+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountIdentifyScenarioVersionsEndpoint` and `IdentifyScenarioVersion` resource
 - Add `AccountSettings.identify`
 - Add multiple fields for `IdentifyScenarioVersion`
+- Add `IdentifyScenariosEndpoint.info` endpoint
 
 ### Changed
 - Change `Identification.envelope` from string to EnvelopeInfo